added one NPC and dialogue, roguelike view of dungeons and corridors

+++ src/tile/TileManager.java
@@ -90,14 +90,14 @@ public class TileManager {
 
     	
     	while (worldCol < gp.maxWorldCol && worldRow < gp.maxWorldRow) {
-    	    int index = dungeon.getRoomIndexAt(gp.player.worldX/gp.tileSize, gp.player.worldY/gp.tileSize);
+    	    int index = dungeon.getRoomIndexAt((int)gp.player.worldX/gp.tileSize, (int)gp.player.worldY/gp.tileSize);
     	    int worldX = worldCol * gp.tileSize;
     	    int worldY = worldRow * gp.tileSize;
-    	    int screenX = worldX - gp.player.worldX + gp.player.screenX;
-    	    int screenY = worldY - gp.player.worldY + gp.player.screenY;
+    	    int screenX = worldX - (int)gp.player.worldX + gp.player.screenX;
+    	    int screenY = worldY - (int)gp.player.worldY + gp.player.screenY;
 
-    	    if ((index != -1 && isWithinBounds(worldCol, worldRow, worldX, worldY, gp.player.worldX, gp.player.worldY, gp.player.screenX, gp.player.screenY, rooms.get(index))) || 
-    	        (index == -1 && isWithinBounds(worldCol, worldRow, worldX, worldY, gp.player.worldX, gp.player.worldY, gp.player.screenX - 480, gp.player.screenY - 355, null))) {
+    	    if ((index != -1 && isWithinBounds(worldCol, worldRow, worldX, worldY, (int)gp.player.worldX, (int)gp.player.worldY, gp.player.screenX, gp.player.screenY, rooms.get(index))) || 
+    	        (index == -1 && isWithinBounds(worldCol, worldRow, worldX, worldY, (int)gp.player.worldX, (int)gp.player.worldY, gp.player.screenX - 480, gp.player.screenY - 355, null))) {
     	        g2.drawImage(dungeonMap[worldRow][worldCol].image, screenX, screenY, null);
     	    }
